+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 [Unreleased]## [0.5.5.2b4]
+### Fixed
+- Small error in ScenarioRunner.insert_outputs_in_db
 
 ## [0.5.5.2b3]- 2023-07-05
 ### Added