+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-[Unreleased]## [0.5.5.2b4]
+[Unreleased]## [0.5.5.2b5]
+
+## [0.5.5.2b4] - 2024-04-10
 ### Added
 - DataManager.extract_solution argument `round_decimals` to allow rounding of solution value
 - DataManager.extract_solution argument `extract_dvar_names` can also be a `Dict[str, str]` where the value if the solumn column name.